refactor: extract cmp for participate button
